Skip sharder probe tests before starting servers
Before it would take approx 33 seconds on my laptop for a test in
probe/test_sharder.py to simply skip, because the checks for skipping
were not made until all services had been started. Now the tests will
skip in approx 23 seconds, which is less frustrating although still
not great.

Change-Id: I048d8ca83555fc225506d85887d61d8a0e7b7588