Add auto_shard option to enable or disable leader behaviour
Add auto_shard option to container-sharder section of container server
config.

When auto_shard is true the sharder daemon will adopt leader behaviour
as before, i.e.  the daemon is the leader for the node index 0 replica
of a container.

When auto_shard is false (the default) the sharder daemon never adopts
leader behaviour.

Change-Id: I28cce3562a683c36ff39e3d99b8c48eaff70f34d