Store epoch in shard ranges table
Storing the epoch of a container in its shard range makes it available
when creating brokers to replicate to that container. This means that
when cleaving, the cleave dbs should have the same epoch as the shard
container to which they are cleaving.

There is no other timestamp in a shard range that remains at the value
of the epoch so it requires a new attribute in ShardRange.

Change-Id: I4b8fec618df9b6df463b0be46ae27ac7486edd97