Remove race between checking for state shrinking and acting upon it
I saw a probe test fail during shrinking because the root shard range
table had both the expanded acceptor in active state and the undeleted
donor in shrinking state. No shrinking activity had actually occurred
on any replica of the donor.

The only way I could see that happening is if the donor sent the
acceptor state to the root prematurely. This could happen if the donor
checked that it was not shrinking, then received the acceptor from the
root, then sent its own range *and the acceptor* back to the root
based on a stale sample of its own state.

This patch also breaks the root updating into a separate method in
order to have targetted unit tests.

Change-Id: I4d3db021388daf12f59648ff06119a30d3444147