Only delete misplaced objects that were moved
Before, we read any misplaced objects from a range of the source db,
replicated them, then removed everything in that range from the source
db. Any objects being merged between the read and the removal would be
lost.

With this patch, two safety measures are added:

First, the broker info is sampled before objects are read. The info is
annotated with the db max_row at that moment. Then, when objects are
removed, only rows up to the max row in the sampled broker info are
removed. Any new rows will be retained and dealt with on the next
cycle.  It is possible that new rows were in fact yielded and
replicated, but it is not certain, so they should not be removed. At
worst they will be replicated again on the next cycle and then
removed.

Second, before removing any rows, a check is made that the broker id
has not changed. If it has then no rows are removed.

Also add unit test for ContainerBroker.remove_objects().

Change-Id: I923c257996eb1bfd3aa17895616750461b3aee59