Stop passing own_shard_range to _misplaced_objects
Move toward a consistent interface to the methods that implement each
phase of sharding i.e. just pass in a ContainerBroker instance.

The change also avoids any potential for a stale copy of
own_shard_range to be used in _misplaced_objects.

Furthermore, the change simplifies testing. It is no longer necessary
to create an own shard range instance simply to pass it to
_misplaced_objects() in tests.

There was previously some efficiency gained from fetching a single
instance of own_shard_range which was passed to _misplaced_objects and
used to evaluate the state of the sharding process. This is traded off
for the advantages described above.

Change-Id: I96ec8a06c859d49823ec97314011694eac87cb16