Only use estimated progress in find_shard_ranges if reliable
Before, ContainerBroker.find_shard_ranges assumed that any existing
shard ranges had same size as the shard_size passed to the current
invocation of the method when estimating progress.  This could lead to
undesirable results if different shard sizes were used for different
invocation, particularly if the shard size increased to make the
progress appear larger than reality.

With this patch the estimate of progress is based on the actual sizes
of previously found shard ranges. For this to be accurate, previously
found ranges must have remained in FOUND state i.e. no shard container
has yet been created and no newer object updates made to the
shards. If this is not true then the progress estimate is considered
unreliable and defaults to zero.

A TODO is removed which suggested that the object_count might be
refreshed before processing each found range upper. This does not seem
necessary because find_shard_ranges() is called after container has
moved to SHARDING state such that the db being scanned is effectively
'read-only'.

Change-Id: Ib74e2efc4e854a2dd95101777bca114a7865fba7