Move misplaced objects when cleaving
Before, when cleaving a shard, only objects that fell within a
destination shard range were cleaved. Any misplaced objects *may* have
been dealt with by a previous invocation of the misplaced objects
handler, but that was not guaranteed.

This patch introduces a further call to misplaced objects specifically
to move any misplaced objects in the database that is being cleaved.

Change-Id: Idd12682529b2da9dc54aa2c4c925d9bb30374a4a