Use db id and max row to key cleaving context
Change the key used to track cleaving progress to be based on the db
id and max row rather than the db hash. The hash might change if the
db is replicated, because replication performs a reclaim of any old
rows, which changes the hash. The max row will not change. However,
max row cannot be assumed to be unique to a particular db so the db id
is also included in the key.

Change-Id: I5c49cf45594b3ce3756450e4b64e41d155a8e99b