Don't skip CLEAVED state when cleaving a root container
The CLEAVED state was introduced in the Related-Change [1].  Sharding
shard containers put their sub-shards to CLEAVED state after each is
cleaved. Sub-shards are moved to ACTIVE when they have all been
cleaved, at which time the sharding shard range is deleted.

While cleaving is in progress, the two states allow the original
ACTIVE shard range to be distinguished from new CLEAVED sub-shards
whose namespaces overlap with the original. Previously, shard ranges
in CLEAVED state were not used when a root container redirected
requests for listings, because they overlap with their 'parent' ACTIVE
shard range.

Conversely, when a root container is cleaving there is no existing
overlapping shard range and so the root moved cleaved shards directly
to the ACTIVE state so that they would be immediately available to use
for listings.

Since [2] the proxy controller can cope with overlapping shard ranges
when building concatenated listings. Both CLEAVED and overlapping
ACTIVE shard ranges can therefore be included when supplying shard
ranges for container listings.  This means that when a root container
knows of a CLEAVED shard range, it is able to provide it to the proxy
controller, allowing it to fetch a listing component directly from the
sub-shard. This also means that the same state transitions can be made
when cleaving both a root or a shard container.

[1] Related-Change: I74df5085f999095a7d38444299abf12038e62d76
[2] Related-Change: Ia4e1f828e164b5eccc6eaebff103ce2328b44bea
Change-Id: I1394bc1db3c438b4bece1ed526a06d5cb6ef2b4a