Only leader actions bump state_timestamp
The state of a shard range can be changed at different times in
different processes, but should always make forwards progress. If one
replica of a shard range has already moved to a state that is ahead
that of another replica then the most advanced state should be
preserved irrespective of the time at which each replica reached their
current state.

The only time a shard range state needs to move 'backwards' is when a
final shard range shrinks back to a root, when the root needs to
change from sharded state to active state. This transition should only
be performed by a single 'leader' process and it is therefore
acceptable to increase the state_timestamp at that moment in time so
that the active state takes precedence over any other more advanced
state with an older timestamp.

In general, the state timestamp can be increased whenever a single
'leader' is singularly responsible for performing a state transition.

Change-Id: I0a38db0a56645fc5c0b0fcf2cbd89a3b5f31cebf