Fix ContainerBroker.get_shard_ranges() sort order
The SQL query used to select shard range rows from a container db
ordered the results by 'lower' and then 'upper'. However, an upper
value of ShardRange.MAX is stored in the db as the empty string, which
sorts before any other upper value. Consequently shard ranges might be
returned in unexpected order. For example, shard ranges with bounds
('', 'k') and ('', '') would be returned in order ('', '') and ('',
'k').

To ensure the correct ordering, the returned shard ranges are now
sorted after the db rows have been converted to ShardRange instances.

Change-Id: Ifca3362b12a2dfb12f05dd812a35878cabd80924