Remove EXPANDING state
The shard range EXPANDING state was introduced in [1] as a means of
preventing an acceptor shard range treating updates from its donor as
misplaced objects during shrinking. The state was needed because the
acceptor namespace was not expanded prior to shrinking starting, so
the donor's objects would appear to be misplaced in the acceptor.

Now that the proxy controller is able to cope with overlapping shard
ranges when building concatenated shard listings, it is possible for
the acceptor namespace to be expanded prior to shrinking and for the
acceptor to temporarily overlap the donor in the root container. This
makes the EXPANDING state unnecessary.

There is no longer any need for the special case that excluded other
shard ranges from updates to root while shrinking since the root now
already knows about the acceptor's expanded namespace.

[1] Related-Change: I43cb6285537a24163bfe1245fe15fe1f1393294b
Change-Id: I47e4ee2123f617d6eb7fdfdce94d8de877a42e5c