Prevent pending file flushing to retiring db
Before, while a container was in sharding state, a container GET
called ContainerBroker.list_objects_iter() on the retiring broker,
with the goal of listing any objects in the retiring
db. list_objects_iter() flushes any pending object updates into the
broker's object table, which in this scenario was the retiring db,
causing the retiring db to be polluted.

The opportunity for pending updates to pollute the retiring db during
a container GET was limited; before calling list_objects_iter() the
container GET method calls the broker get_info_is_deleted() method
which also flushes the pending file, so the vulnerability is only to
updates arriving in the pending file after that point in time.

However, there are other place where the pending file might be flushed
into the *retiring* db, for example as a side effect of calling
get_info() or get_items_since(). Although we generally expect object
updates to be redirected away from a sharding container, this may not
always be the case, for example while a shard range is found but not
yet created.

This patch adds an attribute 'skip_commits' to DatabaseBroker which
when True will prevent the broker ever committing records to its db
via _commit_puts() method.  'skip_commits' is False by default (legacy
behaviour) but is set to True whenever a broker is created for a
retiring db using ContainerBroker.get_brokers().

Note that 'skip_commits' does not prevent objects being directly
merged to the db via a direct call to merge_items().

Change-Id: I268d01a373491c693b793748065d212f9703ffab