Broaden our per-broker exception handling
This way, if we hit an uncaught exception in
_identify_sharding_candidate we won't bomb out of the entire sharding
cycle. Bonus: we'll get more context about which db file caused the
exception.

FWIW, the exception that led to this change was

   OperationalError: no such table: container_stat

coming up from get_shard_ranges as we try to get our own shard range.
Not sure what that's about yet.

Note that we can now have more failures than attempts in the recon
stats.

Change-Id: I717c85c329b72ac880a2f6232852d7821f50a8f2