Copy deleted shard ranges from old db to new db
When creating a fresh epoch db, ensure that all shard ranges are
copied from the old db, including deleted shard ranges. Otherwise an
old, previously deleted shard range may re-appear.

This bug manifested in probe/test_sharder.py/test_shrinking as
follows:

In the first cycle of the test, a container is sharded and then all
shards shrink back to the original root container, meaning that the
final shard range expands to the entire namespace before shrinking and
being deleted.

On the next cycle the container grows again and when searching for
shrink candidates a state version, active, of the old deleted final
shard range would re-appear in the root shard range table, because
there was no record of it being deleted.  This range would appear as
an upper neighbour of a shrinkable donor range and be chosen as
acceptor in preference to the real upper neighbour of the donor.

For example, given two correct ranges '' - 'obj024' and 'obj024' - '',
plus a rogue range '' - '', the rogue range would sort between the two
correct ranges and therefore be chosen as an acceptor for the donor ''
- 'obj024'.

The bug has become apparent since the related change [1] because prior
to [1] the rogue shard range would have sorted before the donor and
therefore not be chosen as an acceptor.

[1] Related-Change: Ifca3362b12a2dfb12f05dd812a35878cabd80924

Change-Id: If29370fc70485e7510fabbd30d9214295c8156be