Fix intermittent failing probetest
When running probetests in a loop I discovered that the
container-servers were not always shutdown cleanly.  When handling the
SIGTERM received by a manager the parent process would occasionally
crash an assertion in the eventlet hub by doing blocking I/O in a signal
handler.  The child process would never receive the forwarded SIGTERM
and continue to accept requests on the listening socket.

By moving all I/O (logging and the associated PipeMutex locking) out of
the signal handler the child processes can reliably shutdown cleanly.

In my samples the test seems to be significantly more stable.

Change-Id: I5436063299f88106353cfa0ecd9a8b0b59c7af25