Support unicode characters in tempauth user names.
Unicode characters in tempauth user names currently prevent these users
from authenticating. This manifests in two places:
1. the call to get_groups() fails to return a set that matches the user
   preventing authentication
2. a new token is generated every time for such users, as the check
   against existing memcache entries always fails to match

The patch fixes the existing behavior with respect to unicode users. As
part of it, the FakeMemcache class had to change to decode the strings
passed to it to mimic the actual memcache client behavior.

Darrell Bishop developed the patch and Timur Alperovich added the tests
for it.

Change-Id: I4ecfae2bca6ffa08ad15e584579ebce707f4628d