Start from last max row when repeating cleaving
When cleaving is completed but the max row of the retiring db has
changed then cleaving is repeated to avoid losing any rows recently
added to the retiring db.  Previously this meant cleaving every row in
the retiring db again, which is unnecessary and inefficient.  Now only
rows that were added since the last cleaving pass began are cleaved on
a subsequent cleaving pass.

This change requires the cleaving context to be keyed by retiring db
id alone. Previously the db max row id was used in the context key, so
that if the max row changed a new context would automatically be
generated. Now, the max row id is included in the context and the
context is explicitly reset when re-cleaving is required. Note that
the context is still specific to a single replica of the db.

In support of this change, the cleaving context is now encapsulated in
a CleavingContext class. The state being maintained is now
sufficiently rich to warrant a class to ease state manipulation and
testing.  This also separates cleaving context concerns out of the
ContainerBroker, which now simply provides a means to store serialized
cleaving context data.

To complete the separation of concerns, the checks made on cleaving
context prior to changing a broker's db state to SHARDED are moved
from the ContainerBroker to the sharder daemon. This also removes the
need for tedious cleaving context setup in tests to simply get a broker
into sharded state.

Change-Id: Ibe3b1e9d69738a4aad23ade7b28fc534eb5786c9