Fix ContainerBroker.empty()
Previously ContainerBroker.empty did not check the contents of the
fresh db for misplaced objects when in sharded state, and did not
consider the content of a retiring db while sharding. Consequently it
was possible to delete a sharding container, or a sharded container
with only misplaced objects.

Own shard range was also erroneously included when considering shard
range object counts.

Now the method checks that both fresh and any retiring db are empty
and that there is zero shard range usage.

Change-Id: I5b77e54c9ce2b6546feacbe856bf944a18f9b29d