Remove shard_cleanups
shard_cleanups was used to delay the deletion of handoff shard dbs
until all cleaving or misplaced object handling had completed for a
given cycle. The one time this was useful was if a destination shard
for misplaced objects overlapped two source shard ranges. Now, a
destination shard handoff db may be created, replicated and deleted
twice, whereas before it was created and deleted once. This
inefficiency is traded off for deleting handoff shard dbs as soon as
they have been replicated in all other cases.

Change-Id: I1e8408451086fe3ba6057ab5779718484a1a2775