Only a leader should set Scan-Done metadata
Before, if a leader found shard ranges and set Scan-Done to True and
then another node had its own shard range state set to sharding
*before* it received any other shard ranges, then the other node
would set Scan-Done to False. This would then prevent any sharding
progress on that node and any other node to which the Scan-Done
metadata was replicated.

Since a leader node is the only node that sets Scan-Done to True it
should also be the only node that sets it to False.

Change-Id: Ie3f1a53902fcaf5439a316e941df90f2a1ae3eeb