Fix fetching own shard range from root during audit
Using the 'includes' param is unreliable when trying to fetch a
specific shard range from the root container because it may match
another overlapping shard range and only the first match will be
returned.  Specifying the shard range name would be specific but is
not currently supported by the container server API, so using marker
and end_marker is a sufficient solution. This has the additional,
unintended, advantage of returning the desired shard range *plus* any
overlapping ranges, which might be useful in the future should the
audit process attempt to correct overlaps.

Change-Id: I4099e53fe5bc91f612ea9905cab724f2298826d5