update requirements to support matching global exclusions
The match for dependency names is case sensitive, so change
"pastedeploy" to "PasteDeploy".

The match for dependencies takes the environment markers into account,
so add a marker for dnspython to restrict its use to python 2.7. There
is a dnspython3 package for use with version 3, and that can be added
separately as part of the python 3 porting work.

Change-Id: Ia2bb2e1846d734609ea9ff37d2ad1ffd37ced950
Depends-On: https://review.openstack.org/560109
Signed-off-by: Doug Hellmann <doug@doughellmann.com>