Improve check for O_TMPFILE support in unit tests
On Python < 3.4, there is no os.O_TMPFILE, so the check for O_TMPFILE
support was always reporting no-support. We now use utils.O_TMPFILE,
which is always defined.

Also, when opening an anonymous tempfile, you must open the directory
(not a file in the directory) with three flags or-ed together:
O_DIRECTORY, O_WRONLY, and O_TMPFILE. Calling open() with a subset of
those flags will receive either EISDIR or EINVAL. The check was trying
to open a file in the temp dir, so was always reporting no-support for
O_TMPFILE. We now correctly open the tempdir.

Change-Id: Iaf48ad9aecea73df4e794818dd15070360bff19f