Fix ContainerBroker.is_deleted() method
Previously is_deleted() failed to consider retiring db objects while
SHARDING nor shard range contents when the container was SHARDED.

Now, the object_count considered in is_deleted() is the same as
that returned in the result of broker.get_info().

In SHARDING or SHARDED states any misplaced objects are *not*
considered in is_deleted.

Change-Id: I62614ea2087d1c6444260823710ea26691b3e894