Enable shard ranges to be fetched from a deleted container
...and also for deleted shard ranges to be fetched.

Shards still need to be able to fetch shard ranges for audit after a
container has been deleted. This patch adds support for a new
X-Backend-Override-Deleted header which if present will cause a
container GET request for shard record types to ignore the deleted
status of the container, unless the container db has been deleted.

To enable deleted shard ranged to be fetched a new
X-Backend-Include-Deleted header is introduced which when present in a
container GET request for shard record types will cause deleted shard
ranges to be included.

Also, return x-backend-record-type header with container GET
responses, and check it is present when shard ranges are
fetched.

Change-Id: I6e01f110e8ff9900a8ed4ea39b7d643e1d5c99a9