Narrow race in _replicate_object
We use an optimistic "lock" to make the state of the broker we're
replicating hasn't changed since we started replicating before deleting.

There's still a small window after the final check for max_row and
delete_db but it's much better than the unguarded window for the
entirety of replication to all nodes.

If the replicator starts into a db while the sharder is still pumping it
full of cleaved objects we'll skip the delete at the end and pick it up
on the next pass.

Co-Authored-By: Alistair Coles <alistairncoles@gmail.com>

Change-Id: I8ac8552a317fe6b492870f4a5fa83f20428e8e6c