Call _sync_other_items from _choose_replication_mode
- move _sync_other_items from db_replicator superclass to the
container replicator subclass, which reduces container specific
changes in the superclass.

- call _sync_other_items before choosing any other sync method, so
that shard ranges are always replicated before objects - this avoids a
potential race when misplaced objects are moved to an acceptor before
the acceptor knows that its shard range has expanded. Before, there
was a risk that another racing process on the acceptor might have
moved the recently merged replicated objects back to where the were
just moved from.

- this change unfortunately means that shard ranges will be
unnecessarily usync'd prior to an rsync_then_merge, but
rsync_then_merge is unlikely for a container with enough objects that
it is sharding or sharded.

Change-Id: I5b32abcd82ac45cf8e928454b6f3d3de22b09886