Make first GET to root return objects or shard ranges as appropriate
Before the first backend container GET request was always for
objects. The proxy then deduced the container's sharding state from
the response headers and if appropriate made a second request for
shard ranges. This was inefficient.

With this patch the first backend container GET request has
X-Backend-Record-Type set to 'auto' which causes the container server
to return shard ranges if it is sharding or sharded, or objects
otherwise. The proxy can deduce which record type it has received from
the response header X-Backend-Record-Type and behave accordingly.

Change-Id: If507d5b5d3e315007bf78f38ebdda8580cbc45cc