Abort rsync_then_merge if destination db has started sharding
If a container db starts to shard, or even completes, while an rsync
then merge is in process to that db, then try hard to abort the
rsync_then_merge.

Check for abort before starting to merge from existing db to the rsync
db, then again before renaming the rsync db.

Change-Id: Ib285efbadb222b7c843fc212e5ae912ccd7b7ead