Remove a won't-fix TODO
When creating shard ranges, a shard will update root with summary of
all newly created sub-shard containers.  There does not seem to be any
need to prevent progress if this summary update fails. The sharder
will atempt to update the root at the end of each visit to a
container, and any successfully created containers will also begin to
notify the root of their existence.

Change-Id: Idaba5384a8fe47dbb072bc0e2b679360691c6cfc