Remove sharding_lock()
This lock was taken two places:

1. when cleaving into a shard broker
2. when removing misplaced objects from a shard broker

These two operations would not execute concurrently if there is just
one sharder process running.  If there were more than one sharder
process, each processing different nodes, then they might visit the
same db when one is visiting the parent container and cleaving to the
shard and one is visiting the shard container and handling misplaced
objects. However, any objects merged by the cleaving process will not
be removed by the misplaced object process because the removal is
limited to a max row count that is sampled at the start of the
misplaced objects handling. It is therefore not necessary to protect
these operations with a lock.

Change-Id: Icb3f9d8843b0fe601a32006adb4dcb29779c8a06