Stop internal client txn_id bleeding into sharder logs
The LogAdapter txn_id is stored in a threading.local object that is a
class attribute, so all instances of the LogAdapter in same thread
will pick up the txn_id set in another instance. That means that after
the internal client has been used to make a request, all subsequent
sharder logs include the request txn_id.

The txn_id might be useful if an error occurs in _fetch_shard_ranges
but is annoying elsewhere.

Change-Id: I4e1961c13be301381907885579d4137cd0a9b16a