Clean up logs a bit
We don't need to log the fact that we created zero ranges, and we don't need
to log at info for *every* container we find.

It *is* handy to write down the time elapsed when we finish cleaving a shard,
though, so I don't have to correlate log lines, parse and diff dates, etc.

Change-Id: I2b2b8a6b8801082f3068ec0f264ab5d4086aa2c8