Always update shard brokers with their own range and sharding metadata
There is no need to conditionally update shard brokers when they are
created. The 'force' flag was a leftover from when own shard range
state was stored in sysmeta and the most recent writer of state would
take precedence.  We consequently needed to avoid updating the
sharding meta-timestamp in sysmeta as a side-effect of getting the
shard broker.

Since [1] the meta-timestamp is managed by the shard range merging
logic so that the most recent time takes precedence rather than the
most recent writer.

[1] Related-Change: Ie65cd4ce0abc98452163f6acdfad3cce5cdd216f

Change-Id: I3f3d13cf1caa684260f2ddfdb731ca941336f319