Don't repeat merges when re-cleaving to existing shard db
Previously, if a shard db was cleaved but failed to replicate, the
cleaving would be repeated i.e. all rows in the shard range would be
merged from the retiring source db to the shard db. This is
unnecessary, and inefficient, if the shard db is still intact from the
previous attempt to cleave and replicate.

Now, once rows have been cleaved, a sync point is stored in the shard
db under the retiring source db id. This sync point is checked before
rows are merged. If the shard db was re-created or replaced by another
out-of-sync copy, the sync point will not be found and all rows will
be merged. If the shard db is unchanged, or is a copy to which the
original shard db has been replicated, the sync point will be found
and no rows will be merged.

Change-Id: Ic9a5e525afcbc265b06d30c8897f7ffa5b3f95a7