Don't modify source ShardRange instance in sharder yield_objects
Updating the lower attribute fo the source shard range *copy* is fine,
but is only a few steps away from a bug if the copy operation was
removed [1].  There is no need to mutate the shard range, just use a
marker variable.

[1] The shard range timestamp would have to also be updated for the
mutated shard range to ever merge, but nevertheless this is not a good
pattern.

Change-Id: I1b9e5e7d20ce628e6f13f50763c72131effc3862