Include devices not being sharded in local devices
Previously the local devices list was restricted to devices that were
included for sharding, which in the extreme case of only one device
would force all handoff shard dbs to be created on same device as the
sharding db.

Now, all devices are included for purposes of chooding a device to
create a shard db.

Also add some unit tests for partition and device filtering.

Change-Id: Id4dfa83103f89a0398ed389a39bfa7c79e2e9a09